@@ -47,93 +47,93 @@
   },
   "homepage": "https://github.com/ngx-formly/ngx-formly#readme",
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/platform-server": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "@ionic/angular": "^4.8.0",
-    "@nguniversal/express-engine": "^7.1.1",
-    "@nguniversal/module-map-ngfactory-loader": "0.0.0",
+    "@angular/animations": "^8.2.10",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.10",
+    "@angular/core": "^8.2.10",
+    "@angular/forms": "^8.2.10",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.10",
+    "@angular/platform-browser-dynamic": "^8.2.10",
+    "@angular/platform-server": "^8.2.10",
+    "@angular/router": "^8.2.10",
+    "@ionic/angular": "^4.11.0",
+    "@nguniversal/express-engine": "^8.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
     "@ngx-loading-bar/core": "^4.2.0",
     "@ngx-loading-bar/router": "^4.2.0",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
-    "@progress/kendo-angular-dropdowns": "^3.5.6",
-    "@progress/kendo-angular-inputs": "^4.2.1",
-    "@progress/kendo-angular-intl": "^1.7.1",
-    "@progress/kendo-angular-l10n": "^1.4.1",
-    "@progress/kendo-theme-default": "^3.7.0",
+    "@progress/kendo-angular-common": "^1.1.0",
+    "@progress/kendo-angular-dropdowns": "^4.2.0",
+    "@progress/kendo-angular-inputs": "^6.3.0",
+    "@progress/kendo-angular-intl": "^2.0.0",
+    "@progress/kendo-angular-l10n": "^2.0.0",
+    "@progress/kendo-angular-popup": "^3.0.1",
+    "@progress/kendo-theme-default": "^4.4.0",
     "@swimlane/ngx-datatable": "^16.0.0",
-    "ag-grid-angular": "^20.2.0",
-    "ag-grid-community": "^20.2.0",
+    "ag-grid-angular": "^21.2.2",
+    "ag-grid-community": "^21.2.2",
     "bootstrap": "^4.3.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.6.9",
+    "core-js": "^3.3.1",
     "express": "^4.15.2",
-    "nativescript-angular": "^7.2.4",
-    "primeng": "^7.1.3",
-    "rxjs": "^6.5.2",
-    "rxjs-compat": "^6.5.2",
-    "tns-core-modules": "^5.4.3",
+    "nativescript-angular": "^8.2.1",
+    "primeng": "^8.0.4",
+    "rxjs": "^6.5.3",
+    "tns-core-modules": "^6.1.1",
     "tslib": "^1.10.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/build-ng-packagr": "^0.13.9",
-    "@angular-devkit/core": "^7.3.9",
-    "@angular-devkit/schematics": "^7.3.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler": "^7.2.15",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@commitlint/cli": "^7.6.1",
-    "@commitlint/config-angular": "^7.6.0",
-    "@commitlint/travis-cli": "^7.6.1",
-    "@schematics/angular": "^7.3.9",
-    "@types/jasmine": "^3.4.0",
-    "@types/jasminewd2": "^2.0.6",
+    "@angular-devkit/build-angular": "^0.803.9",
+    "@angular-devkit/build-ng-packagr": "^0.803.9",
+    "@angular-devkit/core": "^8.3.9",
+    "@angular-devkit/schematics": "^8.3.9",
+    "@angular/cli": "^8.3.9",
+    "@angular/compiler": "^8.2.10",
+    "@angular/compiler-cli": "^8.2.10",
+    "@angular/language-service": "^8.2.10",
+    "@commitlint/cli": "^8.2.0",
+    "@commitlint/config-angular": "^8.2.0",
+    "@commitlint/travis-cli": "^8.2.0",
+    "@schematics/angular": "^8.3.9",
+    "@types/jasmine": "^3.4.4",
+    "@types/jasminewd2": "^2.0.8",
     "@types/json-schema": "^7.0.3",
-    "@types/node": "^10.14.16",
-    "bundlesize": "^0.17.2",
-    "codecov": "^3.5.0",
-    "codelyzer": "^4.5.0",
-    "commitizen": "^3.1.2",
+    "@types/node": "^12.7.12",
+    "bundlesize": "^0.18.0",
+    "codecov": "^3.6.1",
+    "codelyzer": "^5.1.2",
+    "commitizen": "^4.0.3",
     "conventional-github-releaser": "^3.1.3",
     "cypress": "^3.4.1",
-    "cz-conventional-changelog": "^2.1.0",
+    "cz-conventional-changelog": "^3.0.2",
     "github-contributors-list": "^1.2.3",
     "highlight-loader": "^0.7.3",
     "html-loader": "^0.5.5",
-    "husky": "^1.3.1",
+    "husky": "^3.0.9",
     "istanbul-instrumenter-loader": "^3.0.1",
-    "jasmine": "^3.4.0",
-    "jasmine-core": "~3.3.0",
+    "jasmine": "^3.5.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "json-formatter-js": "^2.2.0",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "markdown-loader": "^5.1.0",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^5.6.1",
     "protractor": "^5.4.2",
-    "raw-loader": "^1.0.0",
+    "raw-loader": "^3.1.0",
     "remap-istanbul": "^0.13.0",
-    "standard-version": "^5.0.2",
-    "ts-loader": "^5.2.0",
-    "ts-node": "^8.3.0",
-    "tsickle": "^0.34.3",
-    "tslint": "^5.19.0",
-    "typescript": "~3.2.4",
+    "standard-version": "^7.0.0",
+    "ts-loader": "^6.2.0",
+    "ts-node": "^8.4.1",
+    "tsickle": "^0.37.0",
+    "tslint": "^5.20.0",
+    "typescript": "~3.5.3",
     "webpack-cli": "^3.1.0"
   },
   "config": {
