@@ -90,7 +90,7 @@ describe('ui-bootstrap: Formly Field Select Component', () => {
 
 @Component({selector: 'formly-form-test', template: '', entryComponents: []})
 class TestComponent {
-  @ViewChild(FormlyForm) formlyForm: FormlyForm;
+  @ViewChild(FormlyForm, { static: true }) formlyForm: FormlyForm;
 
   fields = testComponentInputs.fields;
   form = testComponentInputs.form;
