@@ -40,9 +40,9 @@ import { MatDatepickerInput } from '@angular/material/datepicker';
   `,
 })
 export class FormlyDatepickerTypeComponent extends FieldType implements AfterViewInit {
-  @ViewChild(MatInput, <any> { static: true }) formFieldControl!: MatInput;
-  @ViewChild(MatDatepickerInput, <any> { static: true }) datepickerInput!: MatDatepickerInput<any>;
-  @ViewChild('datepickerToggle') datepickerToggle!: TemplateRef<any>;
+  @ViewChild(MatInput, { static: true }) formFieldControl!: MatInput;
+  @ViewChild(MatDatepickerInput, { static: true }) datepickerInput!: MatDatepickerInput<any>;
+  @ViewChild('datepickerToggle', { static: true }) datepickerToggle!: TemplateRef<any>;
 
   defaultOptions = {
     templateOptions: {
