@@ -24,9 +24,9 @@ import { HomeComponent } from './home.component';
     MatMenuModule,
     RouterModule.forRoot([
       { path: '', pathMatch: 'full', component: HomeComponent },
-      { path: 'guide', loadChildren: './guides/guides.module#GuidesModule' },
-      { path: 'ui', loadChildren: './ui/ui.module#UIModule' },
-      { path: 'examples', loadChildren: './examples/examples.module#ExamplesModule' },
+      { path: 'guide', loadChildren: () => import('./guides/guides.module').then(m => m.GuidesModule) },
+      { path: 'ui', loadChildren: () => import('./ui/ui.module').then(m => m.UIModule) },
+      { path: 'examples', loadChildren: () => import('./examples/examples.module').then(m => m.ExamplesModule) },
     ]),
   ],
   bootstrap: [AppComponent],
