@@ -17,8 +17,7 @@ import { isObservable } from 'rxjs';
 export class FormlyField implements OnInit, OnChanges, DoCheck, AfterContentInit, AfterContentChecked, AfterViewInit, AfterViewChecked, OnDestroy {
   @Input() field: FormlyFieldConfig;
 
-  // TODO: remove `any`, once dropping angular `V7` support.
-  @ViewChild('container', <any> {read: ViewContainerRef, static: true }) containerRef: ViewContainerRef;
+  @ViewChild('container', { read: ViewContainerRef, static: true }) containerRef: ViewContainerRef;
   private hostObservers: Function[] = [];
   private componentRefs: any[] = [];
   private hooksObservers: Function[] = [];
