@@ -35,7 +35,7 @@ export class FieldExpressionExtension implements FormlyExtension {
     defineHiddenProp(field, '_expressionProperties', {});
 
     if (field.expressionProperties) {
-      for (const key in field.expressionProperties) {
+      for (const key of Object.keys(field.expressionProperties)) {
         const expressionProperty = field.expressionProperties[key];
 
         if (typeof expressionProperty === 'string' || isFunction(expressionProperty)) {
@@ -174,7 +174,7 @@ export class FieldExpressionExtension implements FormlyExtension {
     let markForCheck = false;
     const expressionProperties = field._expressionProperties;
 
-    for (const key in expressionProperties) {
+    for (const key of Object.keys(expressionProperties)) {
       let expressionValue = evalExpression(expressionProperties[key].expression, { field }, [field.model, field.options.formState, field]);
       if (key === 'templateOptions.disabled') {
         expressionValue = !!expressionValue;
