@@ -22,7 +22,7 @@ export class FieldValidationExtension implements FormlyExtension {
     }
 
     if (field[type]) {
-      for (const validatorName in field[type]) {
+      for (const validatorName of Object.keys(field[type])) {
         validatorName === 'validation'
           ? validators.push(...field[type].validation.map(v => this.wrapNgValidatorFn(field, v)))
           : validators.push(this.wrapNgValidatorFn(field, field[type][validatorName], validatorName))
