@@ -33,7 +33,7 @@ export class FormlyConfig {
   extras: ConfigOption['extras'] = {
     checkExpressionOn: 'changeDetectionCheck',
     lazyRender: false,
-    showError: function(field: FieldType) {
+    showError(field: FieldType) {
       return field.formControl && field.formControl.invalid && (field.formControl.touched || (field.options.parentForm && field.options.parentForm.submitted) || !!(field.field.validation && field.field.validation.show));
     },
   };
@@ -65,7 +65,7 @@ export class FormlyConfig {
       options.forEach((option) => this.setType(option));
     } else {
       if (!this.types[options.name]) {
-        this.types[options.name] = <TypeOption>{ name: options.name };
+        this.types[options.name] = <TypeOption> { name: options.name };
       }
 
       ['component', 'extends', 'defaultOptions', 'wrappers'].forEach(prop => {
@@ -157,7 +157,7 @@ export class FormlyConfig {
 
   setTypeWrapper(type: string, name: string) {
     if (!this.types[type]) {
-      this.types[type] = <TypeOption>{};
+      this.types[type] = <TypeOption> {};
     }
     if (!this.types[type].wrappers) {
       this.types[type].wrappers = [];
@@ -216,9 +216,7 @@ export interface WrapperOption {
   types?: string[];
 }
 
-export interface FieldValidatorFn {
-  (c: AbstractControl, field: FormlyFieldConfig, options?: { [id: string]: any; }): ValidationErrors | null;
-}
+export type FieldValidatorFn = (c: AbstractControl, field: FormlyFieldConfig, options?: { [id: string]: any; }): ValidationErrors | null;
 
 export interface ValidatorOption {
   name: string;
@@ -236,9 +234,7 @@ export interface ValidationMessageOption {
   message: string | ((error: any, field: FormlyFieldConfig) => string | Observable<string>);
 }
 
-export interface ManipulatorWrapper {
-  (f: FormlyFieldConfig): string;
-}
+export type ManipulatorWrapper = (f: FormlyFieldConfig) => string;
 
 export interface TemplateManipulators {
   preWrapper?: ManipulatorWrapper[];
