@@ -3,6 +3,5 @@ import { FieldType } from './field.type';
 import { FormlyFieldConfig } from '../components/formly.field.config';
 
 export abstract class FieldWrapper<F extends FormlyFieldConfig = FormlyFieldConfig> extends FieldType<F> {
-  // TODO: remove `any`, once dropping angular `V7` support.
-  @ViewChild('fieldComponent', <any>{ read: ViewContainerRef, static: false }) fieldComponent: ViewContainerRef;
+  @ViewChild('fieldComponent', { read: ViewContainerRef, static: false }) fieldComponent: ViewContainerRef;
 }
