@@ -31,7 +31,7 @@ import { FieldType } from '@ngx-formly/material/form-field';
   `,
 })
 export class FormlyFieldInput extends FieldType implements OnInit {
-  @ViewChild(MatInput, <any> { static: false }) formFieldControl!: MatInput;
+  @ViewChild(MatInput, { static: false }) formFieldControl!: MatInput;
 
   get type() {
     return this.to.type || 'text';
